requirements.txt
asyncio
tornado
discord.py
pyyaml
requests
sqlalchemy
twine
build